cdp/requirements.txt
Flask
flask_sqlalchemy
flask_migrate
flask-cors
psycopg2
pyjwt
pyotp
gunicorn